rd.py: Basic understanding of hours within days.
That is, you can now specify an hour at which you'd like a reminder to
show up.

Side note that the name "due date" is not really accurate, as this is
rather the timestamp at which we want the reminder to become visible,
which is really going to be strictly before the deadline of the item to
be completed (which is what "due date" means to me).